# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: XMail Gmail analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gmail/response-times:
    get:
      operationId: getResponseTimes
      tags: [gmail]
      summary: Get response time analytics
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            enum: [7, 14, 30]
            default: 7
        - name: accountId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Response time analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTimeAnalytics"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gmail/top-senders:
    get:
      operationId: getTopSenders
      tags: [gmail]
      summary: Get top email senders
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: accountId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Top senders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopSendersResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gmail/weekly-report:
    get:
      operationId: getWeeklyReport
      tags: [gmail]
      summary: Get weekly email report
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Weekly email report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReportResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gmail/dashboard-summary:
    get:
      operationId: getDashboardSummary
      tags: [gmail]
      summary: Get dashboard overview summary
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ResponseTimeEntry:
      type: object
      properties:
        date:
          type: string
        avgResponseMinutes:
          type: number
        emailCount:
          type: integer
      required:
        - date
        - avgResponseMinutes
        - emailCount
    SlowClient:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        avgResponseMinutes:
          type: number
        emailCount:
          type: integer
      required:
        - email
        - name
        - avgResponseMinutes
        - emailCount
    ResponseTimeAnalytics:
      type: object
      properties:
        avgResponseMinutes:
          type: number
        days:
          type: integer
        trend:
          type: array
          items:
            $ref: "#/components/schemas/ResponseTimeEntry"
        slowestClients:
          type: array
          items:
            $ref: "#/components/schemas/SlowClient"
      required:
        - avgResponseMinutes
        - days
        - trend
        - slowestClients
    SenderEntry:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        count:
          type: integer
        category:
          type: string
          enum: [client, team, newsletter, other]
      required:
        - email
        - name
        - count
        - category
    TopSendersResponse:
      type: object
      properties:
        senders:
          type: array
          items:
            $ref: "#/components/schemas/SenderEntry"
        totalEmailsAnalyzed:
          type: integer
      required:
        - senders
        - totalEmailsAnalyzed
...
    DailyVolume:
      type: object
      properties:
        day:
          type: string
        received:
          type: integer
        sent:
          type: integer
      required:
        - day
        - received
        - sent
    WeeklyReportResponse:
      type: object
      properties:
        weekStart:
          type: string
        weekEnd:
          type: string
        totalReceived:
          type: integer
        totalSent:
          type: integer
        busiestDay:
          type: string
        avgResponseMinutes:
          type: number
        dailyVolume:
          type: array
          items:
            $ref: "#/components/schemas/DailyVolume"
      required:
        - weekStart
        - weekEnd
        - totalReceived
        - totalSent
        - busiestDay
        - avgResponseMinutes
        - dailyVolume
    DashboardSummary:
      type: object
      properties:
        totalEmails:
          type: integer
        unreadCount:
          type: integer
        avgResponseMinutes:
          type: number
        inboxScore:
          type: integer
        topSenderEmail:
          type: string
        topSenderName:
          type: string
        weeklyChange:
          type: number
          description: Percentage change vs previous week
      required:
        - totalEmails
        - unreadCount
        - avgResponseMinutes
        - inboxScore
        - topSenderEmail
        - topSenderName
        - weeklyChange
...