# lib/api-zod/src/generated/types/dailyVolume.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XMail Gmail analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyVolume {
  day: string;
  received: number;
  sent: number;
}