# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XMail Gmail analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalEmails: number;
  unreadCount: number;
  avgResponseMinutes: number;
  inboxScore: number;
  topSenderEmail: string;
  topSenderName: string;
  /** Percentage change vs previous week */
  weeklyChange: number;
}