# lib/api-zod/src/generated/types/responseTimeAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XMail Gmail analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { ResponseTimeEntry } from "./responseTimeEntry";
import type { SlowClient } from "./slowClient";

export interface ResponseTimeAnalytics {
  avgResponseMinutes: number;
  days: number;
  trend: ResponseTimeEntry[];
  slowestClients: SlowClient[];
}