# lib/api-zod/src/generated/types/responseTimeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XMail Gmail analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ResponseTimeEntry {
  date: string;
  avgResponseMinutes: number;
  emailCount: number;
}